requirements.txt
streamlit
PyPDF2
google-generativeai
scikit-learn
tiktoken
python-dotenv